requirements.txt
Babel==1.3
blinker==1.3
facebook-sdk==0.4.0
Flask==0.10.1
Flask-Admin==1.0.9
Flask-Babel==0.9
Flask-BabelEx==0.9.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
flask-mongoengine==0.7.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-WTF==0.10.3
google-api-python-client==1.3.1
httplib2==0.9
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mongoengine==0.8.7
oauth2==1.5.211
oauth2client==1.4.5
oauthlib==0.7.2
passlib==1.6.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
pymongo==2.7.2
python-twitter==2.2
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
speaklater==1.3
uritemplate==0.6
Werkzeug==0.9.6
WTForms==1.0.5
python-rake=1.0.6
tweepy=3.5.0